backend_players/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# backend_players/api/v1/players.py: 10
# backend_players/api/v1/train.py: 10
GPUtil == 1.4.0

# backend_players/players/main.py: 2
coloredlogs == 15.0

# backend_players/api/v1/movements.py: 1,2
# backend_players/api/v1/players.py: 1,2
# backend_players/api/v1/train.py: 1,2
# backend_players/main.py: 1,2
fastapi == 0.63.0

# backend_players/players/ArenaOnline.py: 2
# backend_players/players/Coach.py: 8
# backend_players/players/MCTS.py: 4
# backend_players/players/chess/chess_players.py: 1
# backend_players/players/chess/core/board.py: 2
# backend_players/players/chess/keras/NNet.py: 1
# backend_players/players/pit.py: 1
numpy == 1.19.5

# backend_players/schemas/movements.py: 1
# backend_players/schemas/players.py: 1
# backend_players/schemas/train.py: 1
pydantic == 1.7.3

# backend_players/players/ArenaOnline.py: 3
python_socketio == 4.6.1

# backend_players/api/v1/movements.py: 8
# backend_players/api/v1/players.py: 11
# backend_players/api/v1/train.py: 11
# backend_players/players/main.py: 4
requests == 2.25.1

# backend_players/players/chess/keras/ChessNNet.py: 4,5
# backend_players/players/chess/keras/NNet.py: 5
tensorflow == 2.5.0
tensorboard == 2.5.0

# backend_players/players/Arena.py: 3
# backend_players/players/Coach.py: 9
tqdm == 4.61.0

# Uvicorn server
uvicorn == 0.13.4
psycopg2 == 2.8.6
python-multipart == 0.0.5
